Code review comments: - comment LDAP stuff a little more - move new status case classes into model/ so they can be published to Orch - Minor cleanup of HealthMonitor
